Work Completed on 8-28-2018
Formatted the class-wphealthtracker-d3-vitals.php file to conform to WordPress JS standards. Also began work on the Secdon Chart area for the Exercise Stats - complated inclusion of svg and titles - need to complete functionality of hover areas of svg and the labels, to include a readout of data on each muscle group, similar to the Food Items Stats area. Also need to build the actual stats for the stats area of Chart 1.